refactor: move pages to src/pages
